[reactstrap]: Rename utils file
